+++ pkg/server/cloudprovider/errors.go
@@ -1,0 +1,12 @@
+package cloudprovider
+
+import (
+	"github.com/pkg/errors"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+var (
+	ErrNotImplemented       = status.Errorf(codes.Unimplemented, "Operation not implemented for this cloud provider")
+	ErrInvalidCloudProvider = errors.New("invalid cloud provider")
+)