+++ pkg/server/cloudprovider/gce.go
@@ -1,0 +1,9 @@
+package cloudprovider
+
+type gceOperations struct {
+	k8s K8sOperations
+}
+
+func (ops *gceOperations) CreateOrUpdateSSL(appName, cert string, port int) error {
+	return ErrNotImplemented
+}