+++ pkg/server/service/errors.go
@@ -1,0 +1,10 @@
+package service
+
+import (
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+var (
+	ErrNotFound = status.Errorf(codes.NotFound, "Service not found")
+)